Move typechecking code to separate module
